feat(googlelogin): add ErrorBoundary
