+++ packages/@react-oauth/google/src/index.ts
@@ -1,6 +1,6 @@
 export { default as GoogleOAuthProvider } from './GoogleOAuthProvider';
 export { default as GoogleLogin } from './GoogleLogin';
-export type { GoogleLoginProps } from './GoogleLogin';
+export type { GoogleLoginProps } from './GoogleLoginComponent';
 export { default as googleLogout } from './googleLogout';
 export { default as useGoogleLogin } from './hooks/useGoogleLogin';
 export type {
